requirements.txt
blinker==1.9.0
boto3==1.42.60
botocore==1.42.60
click==8.3.1
Deprecated==1.3.1
Flask==3.1.3
Flask-Limiter==4.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
limits==5.8.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.1
